chore: add code style and formart, ts to project and atomic desing structure
