+++ .editorconfig
@@ -1,0 +1,19 @@
+# http://editorconfig.org
+root = true
+[*]
+charset = utf-8
+end_of_line = lf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
+max_line_length = 100
+trim_trailing_whitespace = true
+[*.md]
+max_line_length = 0
+trim_trailing_whitespace = false
+[{Makefile,**.mk}]
+# Use tabs for indentation (Makefiles require tabs)
+indent_style = tab
+[*.scss]
+indent_size = 2
+indent_style = space
