+++ src/old/componets/interfaces/pokemonInterface/Style.jsx
