+++ src/old/componets/interfaces/pokemonStats/Style.jsx
