+++ src/presentation/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const App = () => (
+  <div className="App">
+    <h1>Pokedex</h1>
+    {/* <PokemonView />
+    <PokemonStats attributes={Data} /> */}
+  </div>
+)
+
+export default App
